doc: add doc and helper methods to PDF417

+++ examples/raw.rs
@@ -4,8 +4,8 @@ const WHITE: &str = "\x1B[38;2;255;255;255m█";
 const BLACK: &str = "\x1B[38;2;0;0;0m█";
 
 const PADDING: usize = 4;
-const COLS: u32 = 3;
-const ROWS: u32 = 4;
+const COLS: u8 = 3;
+const ROWS: u8 = 4;
 const LEVEL: u8 = 0;
 
 const INPUT: [u16; 12] = [10, 900, 7 * 30 + 4, 11 * 30 + 11, 14 * 30 + 26, 22 * 30 + 14, 17 * 30 + 11, 3 * 30 + 29, 10 * 30 + 29, 900, 0, 0]; // HELLO WORLD!
@@ -13,8 +13,8 @@ const INPUT: [u16; 12] = [10, 900, 7 * 30 + 4, 11 * 30 + 11, 14 * 30 + 26, 22 * 
 //const INPUT: [u16; 6] = [4, 900, 19 * 30 + 4, 18 * 30 + 19, 0, 0]; // TEST
 //const INPUT: [u16; 20] = [16, 902, 1, 278, 827, 900, 295, 902, 2, 326, 823, 544, 900, 149, 900, 900, 0, 0, 0, 0];
 
-const W: usize = pdf417_width!(COLS) as usize;
-const H: usize = pdf417_height!(ROWS) as usize;
+const W: usize = pdf417_width!(COLS as usize);
+const H: usize = pdf417_height!(ROWS as usize);
 
 fn main() {
     let mut input = INPUT.clone();
