+++ examples/truncated.rs
@@ -4,12 +4,12 @@ const WHITE: &str = "\x1B[38;2;255;255;255m█";
 const BLACK: &str = "\x1B[38;2;0;0;0m█";
 
 const PADDING: usize = 4;
-const COLS: u32 = 3;
-const ROWS: u32 = 6;
+const COLS: u8 = 3;
+const ROWS: u8 = 6;
 const LEVEL: u8 = 1;
 
-const W: usize = pdf417_width!(COLS, 1, 0, true) as usize;
-const H: usize = pdf417_height!(ROWS, 1, 0) as usize;
+const W: usize = pdf417_width!(COLS as usize, 1, 0, true);
+const H: usize = pdf417_height!(ROWS as usize, 1, 0);
 
 fn main() {
     const S: &str = "Truncated PDF417";
@@ -18,7 +18,7 @@ fn main() {
     println!("{data_words}/{}", input.len());
 
     let mut storage = [false; W * H];
-    let pdf417 = PDF417::new(&input, ROWS as u32, COLS as u32, LEVEL)
+    let pdf417 = PDF417::new(&input, ROWS, COLS, LEVEL)
         .truncated(true);
     pdf417.render(&mut storage[..]);
