+++ examples/utf8.rs
@@ -4,13 +4,13 @@ const WHITE: &str = "\x1B[38;2;255;255;255m█";
 const BLACK: &str = "\x1B[38;2;0;0;0m█";
 
 const PADDING: usize = 4;
-const COLS: u32 = 4;
-const ROWS: u32 = 6;
+const COLS: u8 = 4;
+const ROWS: u8 = 6;
 const LEVEL: u8 = 1;
 const SCALE: (u32, u32) = (1, 1);
 
-const W: usize = pdf417_width!(COLS, SCALE.0) as usize;
-const H: usize = pdf417_height!(ROWS, SCALE.1) as usize;
+const W: usize = pdf417_width!(COLS as u32, SCALE.0) as usize;
+const H: usize = pdf417_height!(ROWS as u32, SCALE.1) as usize;
 
 fn main() {
     const S: &str = "💛 ワンピース";
